Move Synchronize to its own file
